fix(windows): fix MSVC build warnings for Release build
This commit fixes multiple MSVC compiler warnings that were causing
build issues on Windows:

- LNK1181: Add WINDOWS_EXPORT_ALL_SYMBOLS to ipb-transport-mqtt
  so the import library (.lib) is generated for the DLL

- C4324: Add pragma to disable intentional cache-line padding warnings
  in headers using alignas() for performance:
  - lockfree_queue.hpp, lockfree_task_queue.hpp, metrics.hpp
  - backpressure.hpp, endpoint.hpp, channel.hpp, message_bus.hpp
  - router.cpp

- C4267: Fix size_t to uint16_t narrowing in data_point.hpp with
  explicit cast

- C4996: Replace deprecated getpid() with platform::get_process_id()
  in test_debug.cpp, and deprecated getenv() with platform::get_env()
  in main.cpp

- C4100/C4101: Fix unreferenced parameters/variables in:
  - paho_backend.cpp, compiled_pattern_cache.cpp
  - mqtt_scoop.cpp, mqtt_sink.cpp

- C4244/C4267: Fix narrowing conversion warnings in:
  - transform_test.cpp, test_backpressure.cpp

+++ core/common/include/ipb/common/data_point.hpp
@@ -432,7 +432,7 @@ public:
             external_address_.reset();
         }
 
-        address_size_ = std::min(address.size(), static_cast<size_t>(UINT16_MAX));
+        address_size_ = static_cast<uint16_t>(std::min(address.size(), static_cast<size_t>(UINT16_MAX)));
 
         if (address_size_ <= MAX_INLINE_ADDRESS) {
             std::memcpy(inline_address_, address.data(), address_size_);
